backend/pyproject.toml
[tool.poetry]
name = "AWattPrice Backend"
version = "2.1"
description = "Backend for the AWattPrice iOS app."
authors = ["Léon Becker <lb@alien8.de>", "Frank Becker <fb@alien8.de>"]
homepage = "https://awattprice.space8.me"
repository = "https://www.github.com/sp4c38/AWattPrice"
# Add here all kinds of additional classifiers as defined under
# https://pypi.python.org/pypi?%3Aaction=list_classifiers
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Intended Audience :: Other Audience",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python",
]
packages = [
    { include = "awattprice", from = "src" },
    { include = "awattprice_notifications", from = "src" },
    { include = "awattprice_helpers", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.63.0"
python-box = "^5.3.0"
loguru = "^0.5.3"
liteconfig = "^1.0.0"
httpx = {extras = ["http2"], version = "^0.18.2"}
arrow = "^1.1.0"
filelock = "^3.0.12"
aiofile = "^3.5.0"
SQLAlchemy = "^1.4.15"
aiosqlite = "^0.17.0"
jsonschema = "^3.2.0"
PyJWT = "^2.1.0"
cryptography = "^3.4.7"
tenacity = "^8.0.1"
gunicorn = "^20.1.0"
uvloop = "^0.15.3"
uvicorn = {extras = ["standard"], version = "^0.14.0"}

[tool.poetry.dev-dependencies]
ipython = "^7.22.0"
black = "^21.4b2"
nest-asyncio = "^1.5.1"
pdoc3 = "^0.9.2"
SQLAlchemy = {extras = ["mypy"], version = "^1.4.17"}
pyflakes = "^2.4.0"

[tool.black]
line-length = 115

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
